rename: `remote_file_system` module
